+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -20,8 +20,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 4;
-constexpr std::uint16_t translation_table_mapping_table_size = 341;
-constexpr std::size_t translation_table_string_table_size = 53526;
+constexpr std::uint16_t translation_table_mapping_table_size = 340;
+constexpr std::size_t translation_table_string_table_size = 53469;
 constexpr std::size_t translation_table_locale_table_size = 29;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -85,7 +85,6 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "TypeScript non-null assertion is not allowed on parameters"sv,
           "TypeScript non-null assertions are not allowed in JavaScript"sv,
           "TypeScript optional properties are not allowed in JavaScript code"sv,
-          "TypeScript static blocks are not supported in JavaScript"sv,
           "TypeScript type annotation requires parentheses"sv,
           "TypeScript type annotations are not allowed in JavaScript code"sv,
           "TypeScript type exports are not allowed in JavaScript"sv,